Add files via upload
Let’s start with the description and explanation of the classes in the Shop Application:
Product: This is the base class for all products. It contains common properties like name and price.
PhysicalProduct: This class extends the Product class and represents physical products in the shop. It adds property weight, which is the weight of a physical product in kilograms.
DigitalProduct: This class also extends the Product class and represents digital products in the shop. It adds property size, which is the size of a digital product in MBs.
Shop: This class represents the shop itself. It contains a list of products that the shop sells and a reference to the warehouse where the products are stored.
Warehouse: This class represents the warehouse where the products are stored. It contains a map that associates each product with its quantity in the warehouse.
ShoppingCart: This class represents a shopping cart for a customer. It contains a list of products that the customer intends to purchase.
Customer: This class represents a customer of the shop. It contains the customer’s shopping cart.

Now, let’s discuss the methods for each class:
Product: This class might have methods like getName(), getPrice(), etc.
PhysicalProduct and DigitalProduct: These classes might have additional methods specific to them, like getShippingWeight() for PhysicalProduct and getDownloadSize() for DigitalProduct.
Shop: This class might have methods like addProduct(Product product), removeProduct(Product product), getProduct(String name), etc.
Warehouse: This class might have methods like addProduct(Product product, int quantity), removeProduct(Product product, int quantity), getProductQuantity(Product product), etc.
ShoppingCart: This class might have methods like addProduct(Product product), removeProduct(Product product), getTotalPrice(), etc.
Customer: This class might have methods like purchase(), addProductToCart(Product product), removeProductFromCart(Product product), etc.